LifeOsBackend/requirements.txt
# Life OS Backend — Requirements
# Phase 1 (core CRUD) + Auth (JWT + bcrypt)

# ── Core Framework ────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
starlette
python-multipart          # Required for OAuth2PasswordRequestForm

# ── Database ──────────────────────────────────────────────────────────────────
motor
pymongo
dnspython

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic[email]           # [email] pulls in email-validator for EmailStr
pydantic_core
annotated-types
typing_extensions

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography] # JWT encode / decode
passlib[bcrypt]           # Password hashing

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv

# ── Async HTTP (used by motor internally, handy for tests) ────────────────────
httpx
anyio

# ── Development & Testing ─────────────────────────────────────────────────────
pytest
pytest-asyncio
requests